search in different file
