+++ src/components/Header/Header.jsx
@@ -1,35 +1,20 @@
-import React, { useState } from "react";
+import React from "react";
 import "@components/Header/header.scss";
 import { Link, useNavigate } from "react-router-dom";
-import SearchIcon from "@assets/images/icons/search.svg";
 import LogoutIcon from "@assets/images/icons/Frame.svg";
 import { useDispatch, useSelector } from "react-redux";
 import { logoutUser } from "@actions/authActions";
-import {
-  searchBlogPosts,
-  getBlogPostsFromLocalStorage,
-} from "@actions/blogActions";
 import Button from "@components/Button/Button";
+import Search from "@components/Search/Search";
 
 const Header = () => {
   const navigate = useNavigate();
   const dispatch = useDispatch();
   const loggedInUser = useSelector((state) => state.auth.loggedInUser);
-  const [searchQuery, setSearchQuery] = useState("");
 
   const handleLogout = () => {
     dispatch(logoutUser());
     navigate("/login");
-  };
-
-  const handleSearch = (event) => {
-    const query = event.target.value;
-    setSearchQuery(query);
-    if (query.trim() === "") {
-      dispatch(getBlogPostsFromLocalStorage());
-    } else {
-      dispatch(searchBlogPosts(query));
-    }
   };
 
   return (
@@ -38,15 +23,7 @@ const Header = () => {
         <Link to="/" className="navbar__logo">
           WellBlog
         </Link>
-        <div className="navbar__search">
-          <input
-            type="search"
-            placeholder="Search"
-            value={searchQuery}
-            onChange={handleSearch}
-          />
-          <img src={SearchIcon} alt="search" />
-        </div>
+        <Search />
         {loggedInUser ? (
           <div className="navbar__menu">
             Welcome
