+++ src/components/Header/header.scss
@@ -4,6 +4,7 @@
   align-items: center;
   background-color: $primary-color;
   display: flex;
+  flex-wrap: wrap;
   justify-content: space-between;
   padding: 1rem 0;
 
@@ -12,25 +13,6 @@
     font-size: 1.5rem;
     font-weight: $font-weight-bold;
     text-decoration: none;
-  }
-
-  &__search {
-    cursor: pointer;
-    display: flex;
-    flex-grow: 0.1;
-    position: relative;
-    img {
-      position: absolute;
-      right: 0.625rem;
-      top: 0.625rem;
-    }
-    input {
-      background-color: $secondary-color;
-      border-radius: 0.2rem;
-      border: none;
-      padding: 0.6rem 1.8rem 0.6rem;
-      width: 100%;
-    }
   }
 
   &__user-name {
