+++ src/components/NotFound/notFound.scss
@@ -1,9 +1,17 @@
 @import "@styles/variables";
 
 .not-found {
-  font-size: 5rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  font-size: 3rem;
   left: 50%;
   position: fixed;
-  top: 50%;
+  top: 65%;
   transform: translate(-50%, -50%);
+  img {
+    height: 18rem;
+    width: 100%;
+  }
 }
