+++ src/components/Search/search.scss
@@ -1,0 +1,20 @@
+@import "@styles/variables";
+
+.navbar__search {
+  cursor: pointer;
+  display: flex;
+  flex-grow: 0.1;
+  position: relative;
+  img {
+    position: absolute;
+    right: 0.625rem;
+    top: 0.625rem;
+  }
+  input {
+    background-color: $secondary-color;
+    border-radius: 0.2rem;
+    border: none;
+    padding: 0.6rem 1.8rem 0.6rem;
+    width: 100%;
+  }
+}
